rename amounts to details
